purchased item successfully moves into current pantry stock
